requirements.txt
-i https://pypi.org/simple
ansi2html==1.9.1; python_version >= '3.7'
asttokens==2.4.1
blinker==1.7.0; python_version >= '3.8'
certifi==2023.11.17; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
comm==0.2.1; python_version >= '3.8'
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.0; python_version >= '3.8'
decorator==5.1.1; python_version >= '3.5'
et-xmlfile==1.1.0; python_version >= '3.6'
exceptiongroup==1.2.0; python_version < '3.11'
executing==2.0.1; python_version >= '3.5'
flask==3.0.0; python_version >= '3.8'
idna==3.6; python_version >= '3.5'
importlib-metadata==7.0.1; python_version >= '3.7'
ipykernel==6.28.0
ipython==8.20.0; python_version >= '3.10'
itsdangerous==2.1.2; python_version >= '3.7'
jedi==0.19.1; python_version >= '3.6'
jinja2==3.1.2; python_version >= '3.7'
jupyter-client==8.6.0; python_version >= '3.8'
jupyter-core==5.7.1; python_version >= '3.8'
markupsafe==2.1.3; python_version >= '3.7'
matplotlib-inline==0.1.6; python_version >= '3.5'
nest-asyncio==1.5.8; python_version >= '3.5'
numpy==1.26.3; python_version < '3.11'
openpyxl==3.1.2
packaging==23.2; python_version >= '3.7'
pandas==2.1.4
parso==0.8.3; python_version >= '3.6'
pexpect==4.9.0; sys_platform != 'win32'
platformdirs==4.1.0; python_version >= '3.8'
plotly==5.18.0; python_version >= '3.6'
prompt-toolkit==3.0.43; python_full_version >= '3.7.0'
psutil==5.9.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.17.2; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.2; python_version >= '3.6'
requests==2.31.0
retrying==1.3.4
setuptools==69.0.3; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
stack-data==0.6.3
tenacity==8.2.3; python_version >= '3.7'
tornado==6.4; python_version >= '3.8'
traitlets==5.14.1; python_version >= '3.8'
typing-extensions==4.9.0; python_version >= '3.8'
tzdata==2023.4; python_version >= '2'
urllib3==2.1.0; python_version >= '3.8'
wcwidth==0.2.13
werkzeug==3.0.1; python_version >= '3.8'
zipp==3.17.0; python_version >= '3.8'
